Enhanced main.tex
I remake the preamble entirely to move from the class `book` to the class `scrbook` belonging to the [KOMA Script bundle](https://www.ctan.org/pkg/koma-script). 

I added comments to all the lines important to understand what the code does and to justify in that way the pertinence of the use of the code I propose.